Add files via upload
In 2021, I prepared this digital interactive map while working on a project in collaboration with UNICEF Bangladesh. For developing this map, I used ArcGIS for geospatial analysis and georeferencing purposes. To make it live, interactive, and digitally accessible for all, I used Google Maps to host the map. I used Google Earth to process the shapefiles and features created in ArcGIS to make them compatible with Google Maps before finally hosting and going live.